Add search tasks functionality, rename packages
