+++ src/main/java/bruh/command/ExitCommand.java
@@ -1,0 +1,17 @@
+package bruh.command;
+
+import bruh.storage.Storage;
+import bruh.tasklist.TaskList;
+import bruh.ui.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+
+    @Override
+    public String runAndGenerateDescription(TaskList taskList, Ui ui, Storage storage) {
+        return "Bye. Hope to see you again soon!";
+    }
+}
