+++ src/main/java/bruh/command/ListTasksCommand.java
@@ -1,0 +1,12 @@
+package bruh.command;
+
+import bruh.storage.Storage;
+import bruh.tasklist.TaskList;
+import bruh.ui.Ui;
+
+public class ListTasksCommand extends Command {
+    @Override
+    public String runAndGenerateDescription(TaskList taskList, Ui ui, Storage storage) {
+        return taskList.listTasks();
+    }
+}
