+++ src/main/java/bruh/exception/InvalidArgumentException.java
@@ -1,0 +1,7 @@
+package bruh.exception;
+
+public class InvalidArgumentException extends DukeException {
+    public InvalidArgumentException() {
+        super("Sorry, but I'm not sure what that means :(");
+    }
+}
