+++ src/main/java/bruh/exception/MissingArgumentException.java
@@ -1,0 +1,7 @@
+package bruh.exception;
+
+public class MissingArgumentException extends DukeException {
+    public MissingArgumentException(String description) {
+        super(description);
+    }
+}
