+++ src/main/java/bruh/task/Todo.java
@@ -1,0 +1,7 @@
+package bruh.task;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description, 'T');
+    }
+}
