--- src/main/java/duke/command/CommandStub.java
@@ -1,18 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-public class CommandStub extends Command {
-
-    @Override
-    public String runAndGenerateDescription(TaskList taskList, Ui ui, Storage storage) {
-        return "Test command run";
-    }
-
-    @Override
-    public String getDescription() {
-        return "Test command description";
-    }
-}
