--- src/main/java/duke/command/ExitCommand.java
@@ -1,17 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-public class ExitCommand extends Command {
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-
-    @Override
-    public String runAndGenerateDescription(TaskList taskList, Ui ui, Storage storage) {
-        return "Bye. Hope to see you again soon!";
-    }
-}
