--- src/main/java/duke/exception/DukeException.java
@@ -1,7 +1,0 @@
-package duke.exception;
-
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
