--- src/main/java/duke/exception/InvalidArgumentException.java
@@ -1,7 +1,0 @@
-package duke.exception;
-
-public class InvalidArgumentException extends DukeException {
-    public InvalidArgumentException() {
-        super("Sorry, but I'm not sure what that means :(");
-    }
-}
