--- src/main/java/duke/exception/MissingArgumentException.java
@@ -1,7 +1,0 @@
-package duke.exception;
-
-public class MissingArgumentException extends DukeException {
-    public MissingArgumentException(String description) {
-        super(description);
-    }
-}
