--- src/main/java/duke/task/Deadline.java
@@ -1,12 +1,0 @@
-package duke.task;
-
-public class Deadline extends TimedTask {
-    public Deadline(String description, String deadline) {
-        super(description, 'D', deadline);
-    }
-
-    @Override
-    public String toString() {
-        return String.format("%s (by: %s)", super.toString(), getDateTimeDesc());
-    }
-}
