--- src/main/java/duke/task/TaskStub.java
@@ -1,5 +1,0 @@
-package duke.task;
-
-public class TaskStub {
-
-}
