--- src/main/java/duke/task/Todo.java
@@ -1,7 +1,0 @@
-package duke.task;
-
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description, 'T');
-    }
-}
